+++ packages/oauth/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.3.0
+
+- [Breaking] Rename type `GetUserType` to `LuciaUser`; remove `GetCreateUserAttributesType`
+- `userAttributes` params for `createUser()` is optional if `Lucia.UserAttributes` is empty
+- Make `Buffer` dependency optional
+
 ## 0.2.7
 
 - Fix type issues with `existingUser` and `createUser()` for `validateCallback()`